feat: staticConfigの構成を整理し、breakおよびbreakDkの設定を新しいファイルに分割しました
