+++ src/staticConfig/index.ts
@@ -1,0 +1,10 @@
+import { breakConfig } from './break'
+import { breakDkConfig } from './breakDk'
+
+export const staticConfig = {
+  break: breakConfig,
+  breakDk: breakDkConfig,
+} as const
+
+export { buildPage3Images } from './shared'
+export type { TrackImageInsert, TrackImageInserts } from './shared'
